+++ extensions/github/src/components/RepositoryActions.tsx
@@ -116,7 +116,7 @@ export default function RepositoryActions({ repository, mutateList, onVisit }: R
 
         {repository.viewerHasStarred ? (
           <Action
-            title="Remove Star from Repository"
+            title="Remove Star From Repository"
             icon={Icon.StarDisabled}
             onAction={removeStar}
             shortcut={{ modifiers: ["cmd", "shift"], key: "f" }}