+++ extensions/scaleway/src/products/webhosting/queries/index.ts
@@ -1,0 +1,1 @@
+export { useAllRegionHostingsQuery } from './useHostingsQuery'