+++ extensions/lastpass/CHANGELOG.md
@@ -1,5 +1,13 @@
 # LastPass Changelog
 
+## [Enhancement] - 2023-03-15
+
+- Add ability to search by ID
+
+## [Enhancement] - 2023-03-13
+
+- Optimized the code
+
 ## [Added sync option] - 2023-03-07
 
 - Added new option to select how often the LastPass vault is synced